requirements.txt
langgraph==0.2.28
langchain==0.3.1
langchain-groq==0.2.0
langchain-community==0.3.1
tavily-python==0.5.0
python-dotenv==1.0.0
streamlit==1.39.0
pydantic==2.9.2
typing-extensions==4.12.2
